Added script for importing data
Auxiliary function in
